requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Faker==19.3.1
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
uuid==1.30
Werkzeug==2.3.6
zipp==3.16.2
